Move equals() from Float[Sum|Min|Max]AggregatorFactory to SimpleFloat...
Addresses #4671
