+++ processing/src/main/java/io/druid/query/aggregation/FloatMaxAggregatorFactory.java
@@ -29,7 +29,6 @@ import io.druid.segment.ColumnSelectorFactory;
 import java.nio.ByteBuffer;
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 
 /**
  */
@@ -123,30 +122,4 @@ public class FloatMaxAggregatorFactory extends SimpleFloatAggregatorFactory
            ", name='" + name + '\'' +
            '}';
   }
-
-  @Override
-  public boolean equals(Object o)
-  {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-
-    FloatMaxAggregatorFactory that = (FloatMaxAggregatorFactory) o;
-
-    if (!Objects.equals(fieldName, that.fieldName)) {
-      return false;
-    }
-    if (!Objects.equals(expression, that.expression)) {
-      return false;
-    }
-    if (!Objects.equals(name, that.name)) {
-      return false;
-    }
-
-    return true;
-  }
-
 }
