+++ processing/src/main/java/io/druid/query/aggregation/FloatMinAggregatorFactory.java
@@ -29,7 +29,6 @@ import io.druid.segment.ColumnSelectorFactory;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Objects;
 
 /**
  */
@@ -130,30 +129,4 @@ public class FloatMinAggregatorFactory extends SimpleFloatAggregatorFactory
            ", name='" + name + '\'' +
            '}';
   }
-
-  @Override
-  public boolean equals(Object o)
-  {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-
-    FloatMinAggregatorFactory that = (FloatMinAggregatorFactory) o;
-
-    if (!Objects.equals(fieldName, that.fieldName)) {
-      return false;
-    }
-    if (!Objects.equals(expression, that.expression)) {
-      return false;
-    }
-    if (!Objects.equals(name, that.name)) {
-      return false;
-    }
-
-    return true;
-  }
-
 }
